+++ components/webflow/actions/publish-site/publish-site.mjs
@@ -4,7 +4,7 @@ export default {
   key: "webflow-publish-site",
   name: "Publish Site",
   description: "Get a site in a specific domain. [See the docs here](https://developers.webflow.com/#publish-site)",
-  version: "0.0.4",
+  version: "1.0.1",
   type: "action",
   props: {
     webflow,