+++ packages/connect-react/src/components/InternalField.tsx
@@ -1,10 +1,8 @@
+import type { ConfigurableProp } from "@pipedream/sdk";
 import { FormFieldContext } from "../hooks/form-field-context";
 import { useFormContext } from "../hooks/form-context";
 import { Field } from "./Field";
 import { useApp } from "../hooks/use-app";
-import {
-  AppInfo, ConfigurableProp,
-} from "@pipedream/sdk";
 
 type FieldInternalProps<T extends ConfigurableProp> = {
   prop: T;
@@ -43,7 +41,7 @@ export function InternalField<T extends ConfigurableProp>({
       setConfiguredProp(idx, value);
     },
     extra: {
-      app: app as AppInfo | undefined, // XXX fix ts
+      app, // XXX fix ts
     },
   };