+++ packages/sdk/jest.config.js
@@ -15,10 +15,13 @@ module.exports = {
     "ts",
     "js",
   ],
+  moduleNameMapper: {
+    "^(.+)\\.js$": "$1",
+  },
   transform: {
     // '^.+\\.[tj]sx?$' to process ts,js,tsx,jsx with `ts-jest`
     // '^.+\\.m?[tj]sx?$' to process ts,js,tsx,jsx,mts,mjs,mtsx,mjsx with `ts-jest`
-    "^.+\\.tsx?$": [
+    "^.+\\.[jt]sx?$": [
       "ts-jest",
       {
         tsconfig: "tsconfig.node.json",