+++ packages/sdk/CHANGELOG.md
@@ -1,6 +1,25 @@
 <!-- markdownlint-disable MD024 -->
 # Changelog
 
+## [1.0.10] - 2024-12-04
+
+### Changed
+
+- Handle correct casing of `stringOptions` in configure prop response
+
+## [1.0.9] - 2024-12-04
+
+### Added
+
+- `triggerDeploy` preview API
+- `client.version` and `x-pd-sdk-version` header
+
+## [1.0.8] - 2024-11-29
+
+### Changed
+
+- Fix fallback WebSocket import (for new components API)
+
 ## [1.0.7] - 2024-11-21
 
 ### Changed